requirements.txt
# Database and API
qdrant-client>=1.12.0
openai>=1.0.0
python-dotenv

# Image Processing
pillow

# AI and Embeddings
torch
transformers>=4.48.0
accelerate
einops

# For faster inference
sentencepiece